Move `useValueSelectorAsync` to its own file
